+++ CHANGELOG.md
@@ -80,7 +80,7 @@
     * EDIT: Once again I prematurely released a patch.
 
 ### v1.0.10
-* Fixed onStep/onDraw callback lag for real.
+* Optimized onStep/onDraw callback lag for real.
 * Added 2 more projectiles to the Instance.projectiles table.
 * Fixed literal memory leak with Buff.find when calling it a lot.
 * RMT will now run its own __initialize before all other mods (not sure why this wasn't already the case).